### 2024-01-06 #6185 @editorialbot
[PRE REVIEW]: PyProximal - scalable convex optimization in Python
**Submitting author:** @mrava87 (<a href="http://orcid.org/0000-0003-0020-2721">Matteo Ravasi</a>)
**Repository:** http://github.com/pylops/pyproximal/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.7.0
