### 2024-01-15 #6185 @mrava87
@arfon, I see you have been assigned as Managing EiC. Since this is my first time submitting to JOSS, am I expected to do anything at this point or I just need to wait for an Editor to be assigned to my manuscript?
