### 2024-01-17 d04409d
fix: use cg instead of lsqr in AffineSet
