### 2024-01-17 45e9946
minor: small change to Hankel class def
