### 2024-01-17 03556ed
fix: allow ProximalGradient to have iteration-dependant epsg
