### 2024-01-26 #6185 @sappelhoff
Hi @mrava87, I will be editing your submission to JOSS. Do you have suggestions for potential reviewers?
