### 2024-01-26 #6185 @mrava87
Great :) @sappelhoff, nice to e-meet you!

I had a look at the `Five most similar historical JOSS papers` but I am not so convinced they are that similar to our work/library. Perhaps `c-lasso` is the most affine, so I would suggest one of their devs `Leo-Simpson`.

There exists a few libraries similar to pyproximal (as we mention in our doc), so based on that, here are few additional reccomended reviewers who contributed to those libraries:
- `nirum`
- `matthieumeo`
- `pmelchior`

PS: I gave you GitHub usernames but I am explicitely not tagging them, so in case you don't select them, they won't know :)
