### 2024-02-05 #6185 @sappelhoff
@Leo-Simpson & @pmelchior & @matthieumeo I am giving this a little bump up your notifications list: I'd still be happy to receive a response from you regarding https://github.com/openjournals/joss-reviews/issues/6185#issuecomment-1914304561
