### 2024-02-05 #6185 @sappelhoff
👋 @nirum & @nwu63 & @mancellin, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I will also try to send you a more formal invitation via email.
