### 2024-02-06 #6185 @sappelhoff
Hi @Leo-Simpson thanks for being a reviewer, there are plenty of aspects to review for a software and you do not need to be an expert in *all* of these aspects. That's also why we usually have two to three reviewers (covering more areas of expertise). I appreciate you taking this on.

>  How much time do we have (roughly) to proceed to the review?

At JOSS we aim to have the review process completed withing 4-6 weeks (unless a reviewer explicitly informs me that they can only start later, or for other reasons)
