### 2024-02-06 #6185 @matthieumeo
Hi @sappelhoff  and all ! Very sorry for the late reply, I was trying to find a potential replacement. Unfortunately I have left academia and my new job leaves me very little time for reviewing activities. It seems you already have enough reviewers however if needed, I can recommend a close collaborator of mine and co-creator of the Pyxu framework, that provides similar functionalities as pyproximal ! Let me know.
