### 2024-02-22 PyLops/pyproximal#161 @nwu63
Update repo link in documentation
This is part of the review at openjournals/joss-reviews#6326

The installation documentation mentions
```
git clone https://github.com/mrava87/pyproximal.git
```

While GitHub redirects links it may be helpful to update this to reflect the new ownership.
