### 2024-02-22 PyLops/pyproximal#162 @nwu63
Missing description of the Rosenbrock function in documentation
This is part of the review at openjournals/joss-reviews#6326.

On [this page](https://pyproximal.readthedocs.io/en/stable/tutorials/nonlinearconstrained.html#sphx-glr-tutorials-nonlinearconstrained-py), the line
> For this example we will use the well-known Rosenbrock function:

should be followed by a mathematical description of the Rosenbrock function, but only a generic function is presented.
