### 2024-02-24 PyLops/pyproximal#161 @mrava87
Thanks @nwu63 for the catch, fixed :)
