### 2024-02-24 6cebb94
doc: Added Rosebrock function definition
