### 2024-02-24 PyLops/pyproximal#162 @mrava87
@nwu63 Good idea, added here https://github.com/PyLops/pyproximal/pull/165, what do you think?
