### 2024-02-26 PyLops/pyproximal#166 @nwu63
Comments on the JOSS paper
For reference: openjournals/joss-reviews#6326

Here are some comments I have regarding the paper itself:
* In the statement of need, it would be helpful to cite a few well-known packages in order to highlight the key distinguishing features of this package. This is already done in the repo README and can probably be reused.
* One DOI is missing -- see [here](https://github.com/openjournals/joss-reviews/issues/6326#issuecomment-1929141719) for the suggested DOI
