### 2024-02-26 #6326 @nwu63
@sappelhoff I recommend this paper for acceptance, pending the following points to be addressed:
* In the statement of need, it would be helpful to cite a few well-known packages in order to highlight the key distinguishing features of this package. This is already done in the repo README and can probably be reused.
* One DOI is missing -- see https://github.com/openjournals/joss-reviews/issues/6326#issuecomment-1929141719 for the suggested DOI

This is tracked at PyLops/pyproximal#166
