### 2024-02-27 PyLops/pyproximal#166 @mrava87
@nwu63 thanks a lot for the feedback. I will shortly update the paper and let you know so you can check :)
