+++ joss/paper.md
@@ -60,7 +60,11 @@ of convex optimization can also benefit from this library in a number of ways wh
 they can easily include their newly developed proximal operators and solvers; second, they can compare these methods 
 with state-of-the-art algorithms already provided in the library.
 
-`PyProximal` heavily relies on and seamlessly integrates with `PyLops` [@Ravasi:2020], a Python library for matrix-free linear algebra 
+Several projects in the Python ecosystem provide implementations of proximal operators and/or algorithms, 
+which present some overlap with those available in `PyProximal`. A (possibly not exhaustive) list of other projects is composed of
+*The Proximity Operator Repository* [@Chierchia], *proxalgs* [@Maheswaranathan], *proxmin* [@Melchior], *ProxImaL* [@Chan], 
+and *pyxu* [@Kashani]. A key common feature of all of the above mentioned packages is to be self-contained; as such, not only proximal operators and solvers
+are provided, but also linear operators that are useful for the applications that the package targets. On the other hand, `PyProximal` heavily relies on and seamlessly integrates with `PyLops` [@Ravasi:2020], a Python library for matrix-free linear algebra 
 and optimization. As such, it can seamlessy handle problems with millions of unknowns and inherits 
 the interchangle CPU/GPU backend of PyLops [@Ravasi:2021]. More specifically, `PyLops` is leveraged in the implementation of proximal operators that require 
 access to linear operators (e.g., numerical derivatives) and/or least-squares solvers 
