+++ joss/paper.bib
@@ -85,16 +85,14 @@
 }
 
 @article{Leblanc:2023,
-    journal = {Second EAGE Subsurface Intelligence Workshop},
     year = 2023,
     author = {O. Leblanc and M. Hofer and S. Sivankutty and H. Rigneault and L. Jacques},
     title = "{Interferometric Lensless Imaging: Rank-one Projections of Image Frequencies with Speckle Illuminations}",
     url = {https://arxiv.org/abs/2306.12698},
-  	archivePrefix = {arXiv},
+  	doi = {10.1109/tci.2024.3359178},
+    archivePrefix = {arXiv},
   	journal = {ArXiv e-prints},
-
 }
-
 
 @article{Ravasi:2021,
     author = {M. Ravasi},
@@ -105,3 +103,39 @@
     doi = {10.3997/2214-4609.2021612003},
     url = {https://www.earthdoc.org/content/papers/10.3997/2214-4609.2021612003}
 }
+
+@online{Chierchia,
+  author = {G. Chierchia and E. Chouzenoux and P. L. Combettes and J.-C. Pesquet},
+  title = {The Proximity Operator Repository},
+  url = {https://proximity-operator.net/},
+}
+
+@online{Maheswaranathan,
+  author = {N. Maheswaranathan and S. Zapp and B. Poole},
+  title = {proxalgs},
+  url = {https://github.com/ganguli-lab/proxalgs/},
+}
+
+@online{Melchior,
+  author = {P. Melchior and F. Moolekamp},
+  title = {proxmin},
+  url = {https://github.com/pmelchior/proxmin/},
+}
+
+@online{Kashani,
+  author = {S. Kashani and M. Simeoni et al.},
+  title = {pyxu},
+  url = {https://github.com/pyxu-org/pyxu/},
+}
+
+@online{Chan,
+  author = {A. Chan and S. Diamond et al.},
+  title = {ProxImaL},
+  url = {https://github.com/comp-imaging/ProxImaL/},
+}
+
+
+
+
+
+
