+++ joss/paper.bib
@@ -104,31 +104,36 @@
     url = {https://www.earthdoc.org/content/papers/10.3997/2214-4609.2021612003}
 }
 
-@online{Chierchia,
-  author = {G. Chierchia and E. Chouzenoux and P. L. Combettes and J.-C. Pesquet},
-  title = {The Proximity Operator Repository},
-  url = {https://proximity-operator.net/},
-}
-
 @online{Maheswaranathan,
+  year = 2019,
   author = {N. Maheswaranathan and S. Zapp and B. Poole},
   title = {proxalgs},
   url = {https://github.com/ganguli-lab/proxalgs/},
 }
 
 @online{Melchior,
+  year = 2022,
   author = {P. Melchior and F. Moolekamp},
   title = {proxmin},
   url = {https://github.com/pmelchior/proxmin/},
 }
 
+@online{Chierchia,
+  year = 2024,
+  author = {G. Chierchia and E. Chouzenoux and P. L. Combettes and J.-C. Pesquet},
+  title = {The Proximity Operator Repository},
+  url = {https://proximity-operator.net/},
+}
+
 @online{Kashani,
+  year = 2024,
   author = {S. Kashani and M. Simeoni et al.},
   title = {pyxu},
   url = {https://github.com/pyxu-org/pyxu/},
 }
 
 @online{Chan,
+  year = 2024,
   author = {A. Chan and S. Diamond et al.},
   title = {ProxImaL},
   url = {https://github.com/comp-imaging/ProxImaL/},
