+++ joss/paper.md
@@ -62,7 +62,7 @@ with state-of-the-art algorithms already provided in the library.
 
 Several projects in the Python ecosystem provide implementations of proximal operators and/or algorithms, 
 which present some overlap with those available in `PyProximal`. A (possibly not exhaustive) list of other projects is composed of
-*The Proximity Operator Repository* [@Chierchia], *proxalgs* [@Maheswaranathan], *proxmin* [@Melchior], *ProxImaL* [@Chan], 
+*proxalgs* [@Maheswaranathan], *proxmin* [@Melchior], *The Proximity Operator Repository* [@Chierchia], *ProxImaL* [@Chan], 
 and *pyxu* [@Kashani]. A key common feature of all of the above mentioned packages is to be self-contained; as such, not only proximal operators and solvers
 are provided, but also linear operators that are useful for the applications that the package targets. On the other hand, `PyProximal` heavily relies on and seamlessly integrates with `PyLops` [@Ravasi:2020], a Python library for matrix-free linear algebra 
 and optimization. As such, it can seamlessy handle problems with millions of unknowns and inherits 
