### 2024-02-29 PyLops/pyproximal#166 @mrava87
@nwu63 thanks again for the suggestions, they are now incorporated in the paper.
