### 2024-03-04 c957605
fix: change print of x[0] to work with cupy
