### 2024-03-11 #6326 @sappelhoff
@Leo-Simpson I haven't heard back from you in 5 weeks now, so I am going to unassign you from the list of reviewers. I'd be happy to welcome you as a reviewer on another project, if the need of JOSS for a reviewer and your own time for volunteering should align again in the future.

@mrava87 we do not need to find another reviewer, as we have two full reviews ready and I judge this situation adequate to proceed.
