### 2024-03-11 #6326 @sappelhoff
@mrava87 yes! In particular, we will now need a tagged release of your software, and an archive of that tagged version (for example on Zenodo, using the GitHub integration). I am going to let the bot create a checklist for us in the next comment.
