### 2024-03-11 #6326 @sappelhoff
@mrava87 please let me know once you finished your items from the list above: https://github.com/openjournals/joss-reviews/issues/6326#issuecomment-1988377528

I will then work on my final items, and then it's (hopefully) finally time to ship it!
