+++ joss/paper.bib
@@ -13,7 +13,8 @@
   	Author = {N. Parikh},
   	Booktitle = {Proximal Algorithms},
   	Publisher = {Foundations and Trends in Optimization},
-  	Year = 2013
+  	Year = 2013,
+  	doi = {doi.org/10.1561/2400000003}
 }
 
 @book{Combettes:2011,
@@ -22,7 +23,8 @@
   	Booktitle = {Fixed-Point Algorithms for Inverse Problems in Science and Engineering.},
   	Publisher = {Springer Optimization and Its Applications},
   	Title = {Proximal splitting methods in signal processing},
-  	Year = 2011
+  	Year = 2011,
+    doi = {10.1007/978-1-4419-9569-8_10},
 }
 
 @article{Boyd:2011,
@@ -111,11 +113,14 @@
   url = {https://github.com/ganguli-lab/proxalgs/},
 }
 
-@online{Melchior,
-  year = 2022,
-  author = {P. Melchior and F. Moolekamp},
-  title = {proxmin},
-  url = {https://github.com/pmelchior/proxmin/},
+@article{Melchior,
+    author = {F. Moolekamp and P. Melchior},
+    title = "{Block-simultaneous direction method of multipliers: a proximal primal-dual splitting algorithm for nonconvex problems with multiple constraints}",
+    journal = {Optimization and Engineering},
+    year = 2018,
+    volume = 19,
+    doi = {10.1007/s11081-018-9380-y},
+    url = {https://link.springer.com/article/10.1007/s11081-018-9380-y}
 }
 
 @online{Chierchia,
@@ -125,18 +130,26 @@
   url = {https://proximity-operator.net/},
 }
 
-@online{Kashani,
-  year = 2024,
-  author = {S. Kashani and M. Simeoni et al.},
-  title = {pyxu},
-  url = {https://github.com/pyxu-org/pyxu/},
+@software{pyxu-framework,
+  author       = {Matthieu Simeoni and
+                  Sepand Kashani and
+                  Joan Rué-Queralt and
+                  Pyxu Developers},
+  title        = {pyxu-org/pyxu: pyxu},
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.4486431},
+  url          = {https://doi.org/10.5281/zenodo.4486431}
 }
 
-@online{Chan,
-  year = 2024,
-  author = {A. Chan and S. Diamond et al.},
-  title = {ProxImaL},
-  url = {https://github.com/comp-imaging/ProxImaL/},
+@article{Heide:2016,
+author = {Heide, Felix and Diamond, Steven and Nie\ss{}ner, Matthias and Ragan-Kelley, Jonathan and Heidrich, Wolfgang and Wetzstein, Gordon},
+title = {ProxImaL: efficient image optimization using proximal algorithms},
+year = {2016},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {35},
+number = {4},
+doi = {10.1145/2897824.2925875},
 }
 
 
