+++ joss/paper.bib
@@ -135,6 +135,7 @@
                   Sepand Kashani and
                   Joan Rué-Queralt and
                   Pyxu Developers},
+  year         = {2024},
   title        = {pyxu-org/pyxu: pyxu},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.4486431},
