+++ joss/paper.bib
@@ -14,7 +14,7 @@
   	Booktitle = {Proximal Algorithms},
   	Publisher = {Foundations and Trends in Optimization},
   	Year = 2013,
-  	doi = {doi.org/10.1561/2400000003}
+  	doi = {10.1561/2400000003}
 }
 
 @book{Combettes:2011,
