### 2024-03-14 #6326 @mrava87
@sappelhoff just to confirm that the paper DOI works and everything looks good :)

But I do not seem to be able myself to close this issue (point 2), I guess you can...

Thanks again everyone, @sappelhoff @nirum and @ewu63!
